clean_and_crop/requirements.txt
# Usage: pip install -r requirements.txt

# base : 
numpy>=1.24
matplotlib>=3.3
opencv-python>=4.1.1
argparse
imutils


# YOLOv5 requirements

gitpython>=3.1.30
pillow>=10.3.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.32.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.2.34  # https://ultralytics.com
pandas>=1.1.4
seaborn>=0.11.0
tf_slim


